version 1.0

+++ __main__.py
@@ -3,8 +3,7 @@ The main module of app
 """
 from postgres import Postgres
 from sqlite import SQLite
-
-from utils import get_inputs
+from utils import get_inputs, select
 
 database = Postgres()
 
@@ -13,24 +12,33 @@ COLUMNS = [('column1', 'varchar(255)'), ('column2', 'varchar(255)')]
 
 database.create(TABLE, [f"{column[0]} {column[1]}" for column in COLUMNS])
 
-data = get_inputs('Insira a', COLUMNS)
-result = database.insert(TABLE, data)
-print(f"Insert {result}")
+OPTIONS = ["SELECT", "INSERT", "UPDATE", "DELETE"]
 
-result = database.select(TABLE)
-print(f"Select {result}")
+while True:
+    INDEX = select(OPTIONS)
 
-where = get_inputs('Selecione a', COLUMNS)
-result = database.select(TABLE, where)
-print(f"Select {result}")
+    if INDEX == 0:
+        break
 
-data = get_inputs('Atualize a', COLUMNS)
-where = get_inputs('Onde a', COLUMNS)
-result = database.update(TABLE, data, where)
-print(f"Update {result}")
+    if INDEX == 1:
+        where = get_inputs('Selecione a', COLUMNS)
+        result = database.select(TABLE, where)
+        print(f"Select {result}")
 
-""" 
-where = get_inputs('Delete onde', COLUMNS)
-result = database.delete(TABLE, where)
-print(f"Delete {result}") 
-"""
+    elif INDEX == 2:
+        data = get_inputs('Insira a', COLUMNS)
+        result = database.insert(TABLE, data)
+        print(f"Insert {result}")
+
+    elif INDEX == 3:
+        data = get_inputs('Atualize a', COLUMNS)
+        where = get_inputs('Onde a', COLUMNS)
+        result = database.update(TABLE, data, where)
+        print(f"Update {result}")
+
+    elif INDEX == 4:
+        where = get_inputs('Delete onde', COLUMNS)
+        result = database.delete(TABLE, where)
+        print(f"Delete {result}")
+
+    input()
